+++ src/main/java/searchengine/controllers/ApiController.java
@@ -8,7 +8,6 @@ import searchengine.services.IndexingService;
 import searchengine.dto.search.SearchResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import searchengine.services.RepositoryService;
 import searchengine.services.SearchService;
 import searchengine.dto.statistics.StatisticsResponse;
 import searchengine.services.StatisticsService;
@@ -23,7 +22,6 @@ public class ApiController {
 	private final SearchService searchService;
 	private final IndexingService indexingService;
 	private final StatisticsService statisticsService;
-	private final RepositoryService repositoryService;
 
 	@GetMapping("/statistics")
 	public ResponseEntity<StatisticsResponse> statistics() {
@@ -53,15 +51,5 @@ public class ApiController {
 			@RequestParam final Integer limit) {
 
 		return ResponseEntity.ok(searchService.getSearchResults(query, site, offset, limit));
-	}
-
-	@GetMapping("/delSite")
-	public void deleteById(@RequestParam final Integer id) {
-		repositoryService.deleteSiteById(id);
-	}
-
-	@GetMapping("/delPage")
-	public void deletePageById(@RequestParam final Integer id) {
-		repositoryService.deletePageById(id);
 	}
 }